:recycle: (Dashboard) extract FavoriteTrainingCard
