+++ src/pages/Home/Dashboard/components/FavoriteTrainingCard/index.ts
@@ -1,0 +1,1 @@
+export { FavoriteTrainingCard } from './FavoriteTrainingCard';
